requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
openai==1.85.0
Pillow==11.2.1
protobuf==6.31.1
pydantic==2.11.5
python-dotenv==1.1.0
google-genai==1.18.0
email_validator==2.2.0
python-multipart==0.0.20
pymongo
celery[redis]
beautifulsoup4
transformers
torch           
pandas
openpyxl
sentence-transformers==2.7.0
scikit-learn
pypdf 
python-docx 
langchain
faiss-cpu
requests